+++ RELEASE_NOTES.md
@@ -1,2 +1,8 @@
+## 0.1.0 - May 5 2017
+* Clean up Paket dependencies
+* Change `IrcRecipient` to `IrcPrefix`
+* Added pattern matching for nickname, user and hostmasks
+* Added unit tests
+
 ### 0.0.1 - April 27 2017
 * Initial release